feat: add Safari parallax background
