+++ src/components/molecules/SafariParallaxBackground/index.tsx
@@ -1,0 +1,1 @@
+export {ParallaxBackground} from './ParallaxBackground'
